+++ logger/docs/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.0.6] - 2023-02-09
 
-- `loggerModule` now embeds `modules.Module` making it registerable
+- `loggerModule` type-checking for `modules.Module`
 
 ## [0.0.0.5] - 2023-02-07